Move source from info_generator file back into main.cpp
