+++ tools/info/CMakeLists.txt
@@ -32,8 +32,6 @@ target_sources(gfxrecon-info
                PRIVATE
                    ${CMAKE_CURRENT_LIST_DIR}/../tool_settings.h
                    ${CMAKE_CURRENT_LIST_DIR}/main.cpp
-                   ${CMAKE_CURRENT_LIST_DIR}/info_generator.h
-                   ${CMAKE_CURRENT_LIST_DIR}/info_generator.cpp
                    ${CMAKE_CURRENT_LIST_DIR}/info_api_generator.h
                    ${CMAKE_CURRENT_LIST_DIR}/info_d3d12_generator.h
                    ${CMAKE_CURRENT_LIST_DIR}/info_d3d12_generator.cpp
